backend/requirements.txt
feedparser>=6.0.11
httpx>=0.27
python-dateutil>=2.9
anthropic>=0.72
supabase>=2.9
python-dotenv>=1.0
tzdata>=2024.1; sys_platform == 'win32'
